+++ src/lib/shapefile.js
@@ -9,7 +9,7 @@ const Cmd = require('./ogr-cmd')
 function createStream (options) {
   const readStream = _()
   const cmd = `ogr2ogr ${Cmd.create('zip', options).join(' ')}`
-  exec(cmd, err => {
+  const ogr = exec(cmd, err => {
     if (err) return readStream.emit('error', new Error('Call to OGR failed'))
     const folder = `${options.path}/${options.name}`
     if (options.metadata) fs.writeFileSync(`${folder}/${options.name}.xml`, options.metadata)
@@ -17,6 +17,7 @@ function createStream (options) {
     .on('error', e => readStream.emit('error', e))
     .pipe(readStream)
   })
+  process.on('SIGTERM', () => ogr.kill())
   return readStream
 }
 
@@ -24,10 +25,11 @@ function createZipStream (input) {
   const zipStream = _()
   const cmd = ['zip', '-rj', `"${input}.zip"`, `"${input}"`, '--exclude=*.json*', '--exclude=*.vrt']
   renameFiles(input)
-  exec(cmd.join(' '), (err, stdout, stderr) => {
+  const zip = exec(cmd.join(' '), (err, stdout, stderr) => {
     if (err) return zipStream.emit('error', new Error('Failed while zipping'))
     fs.createReadStream(`${input}.zip`).pipe(zipStream)
   })
+  process.on('SIGTERM', () => zip.kill())
   return zipStream
 }